Add generic status effect framework + migrate MoreDamage effect
Let's say I want an effect to reduce the cooldown of all food items.
I don't want there to be code in every food class to specially handle
a "reduce cooldown" status effect. Let's define the effect centrally
and automatically apply it to certain items/properties based on tags.

Another benefit: We can use an arbitrary function to compute the final
effect intensity. Ex: custom ramp down based on effect duration
remaining.

Right now, it's kind of boilerplate-y. But we'll iterate on that